backend/requirements.txt
fastapi
uvicorn
tensorflow-cpu
opencv-python-headless
numpy
python-multipart
